This one's pretty destructive:
Rename all of the field types ( geo_point, geo_linestring, geo_polygon ) to
just "geo".  This is to reduce the confusion of chewing up the field type
selector with a whole bunch of nonsense words.  It also gives us the opportunity
to document and explain the various storage types on the field settings page.

Making the available geo_types an API call, so you could theoretically have 3D
types available, and/or stick with "Point" for dumb database storage.

So, there's now 1 field type on the "Add field" form, and the avialable geo
types are on the field settings page. This setting can't be updated.

Naturally, this means the field types have gone away, and you've now got
orphan fields on your content types.  Sorry :(  This is presumably fixable?