Allow CCK to handle field storage "natively" using the Schema API
This means removing over 100 lines of hacky code, and prevents the addition of
another pile of code to address the side-effects.

This code:
  1) Provides a geo API call that returns a field definition that can be used
     by the Schema API for creating new columns.
  2) Removes all of the code in the Geo Field module that forces a field to
     "multiple" and then sneaks up behind CCK and molests its tables.  CCK
     is now free to create and move Geo data with impunity.

STILL TO DO
Because we're adding the columns manually and not with e.g. PostGIS'
AddGeometryColumn() function, we'll need to account for the post-processing
and indexes on the resulting tables.

That's next.