SOME VERY LARGE AND IMPORTANT CHANGES!
1. One of my big TODO items has been to document the OGC specifications for
geospatial data and consistently implement it as an API.  This commit
adds an Interface class for the specification, plus implementations for
MySQL, PostGIS and some strictly PHP-based calculations.

These will eventually supercede some of geo.inc's functions and the files
in the db directory.  As of now, they do not.

This allows us to account for all of the specification's functionality, and
to consistently identify what's supported by the available backends.

2. Another big TODO item has been to track each geospatial dataset and
its metadata in a central table.  My first attempt was in the geo_data
module, but is better managed by the Geo API itself.

This commit adds a tracking table "{geo}" to catalog all geo datasets, and
copies whatever we can find into that table.

3. Adding a handful of API functions to geo.module to access the above
functionality - list handlers, load a handler, list datasets, etc.