Ensuring that all filters and arguments for geo and related geo data use a separate where group called 'geo'.  This will permit us to convert Geo's filters from AND to OR.
For example, you can filter a view based on an event's location OR the location
of a facility that an event is node-referenced to.

Or, you could find your statewide representative whose state overlaps yours OR
a representaive whose district overlaps yours... thus returning all
representatives for a particular user.

And so-forth.